+++ src/components/post-comment.jsx
@@ -10,10 +10,8 @@ import {commentReadmoreConfig} from '../utils/readmore-config';
 
 import CommentLikes from './comment-likes';
 import PieceOfText from './piece-of-text';
-import CommentActionsMobile from './comment-actions-mobile';
 import Expandable from './expandable';
 import UserName from './user-name';
-import TimeDisplay from './time-display';
 
 
 export default class PostComment extends React.Component {
@@ -30,21 +28,6 @@ export default class PostComment extends React.Component {
     this.setState({
       editText: event.target.value
     });
-  }
-
-  openAnsweringComment = (event) => {
-    event.preventDefault();
-    if (this.props.openAnsweringComment && event.button === 0) {
-      const withCtrl = event.ctrlKey || event.metaKey;
-      if (!withCtrl && this.props.hideType) {
-        return;
-      }
-      if (event.ctrlKey || event.metaKey) {
-        this.reply();
-      } else {
-        this.mention();
-      }
-    }
   }
 
   reply = () => {
@@ -214,37 +197,21 @@ export default class PostComment extends React.Component {
     );
   }
 
-  renderTimer() {
-    return <div>
-              <TimeDisplay className="comment-time" timeStamp={+this.props.createdAt} timeAgoInTitle={true}>
-               <a
-                  className={`comment-icon fa ${this.props.omitBubble ? 'feed-comment-dot' : 'fa-comment-o'}`}
-                  id={`comment-${this.props.id}`}
-                  href={`${this.props.entryUrl}#comment-${this.props.id}`}
-                  onClick={this.openAnsweringComment}></a>
-              </TimeDisplay>
-              <CommentActionsMobile omitBubble={this.props.omitBubble}
-                                    likes={this.props.likes}
-                                    forbidLiking={this.props.isEditable}
-                                    hasOwnLike={this.props.hasOwnLike}
-                                    toggleLike={this.toggleLike}
-                                    likesList={this.props.likesList}
-                                    getCommentLikes={this.getCommentLikes}
-                                    reply={this.reply}
-                                    mention={this.mention}/>
-            </div>;
-  }
-
   renderCommentLikes() {
-    if (this.props.hideType || this.props.isEditing) {
+    if (this.props.hideType) {
       return false;
     }
-    return <CommentLikes  likes={this.props.likes}
+    return <CommentLikes  commentId={this.props.id}
+                          entryUrl={this.props.entryUrl}
+                          omitBubble={this.props.omitBubble}
+                          likes={this.props.likes}
                           forbidLiking={this.props.isEditable}
                           hasOwnLike={this.props.hasOwnLike}
                           toggleLike={this.toggleLike}
                           likesList={this.props.likesList}
-                          getCommentLikes={this.getCommentLikes}/>;
+                          getCommentLikes={this.getCommentLikes}
+                          reply={this.reply}
+                          mention={this.mention}/>;
   }
 
   render() {
@@ -257,7 +224,6 @@ export default class PostComment extends React.Component {
 
     return (
       <div className={className} data-author={this.props.isEditing ? '' : this.props.user.username}>
-        {this.renderTimer()}
         {this.renderCommentLikes()}
         {this.renderBody()}
       </div>
