+++ styles/helvetica/app.scss
@@ -24,7 +24,6 @@
 @import "../shared/arrow-span";
 @import "../shared/expandable";
 @import "../shared/comment-likes";
-@import "../shared/comment-actions-mobile";
 @import "../shared/bookmarklet";
 @import "player";
 @import "react-select";
