--- styles/shared/comment-actions-mobile.scss
@@ -1,106 +1,0 @@
-.comment-actions-mobile {
-  @media (min-width: 1025px) {
-    display: none;
-  }
-
-  .icon {
-    color: #ffb914;
-  }
-
-  .feed-comment-dot {
-    height: 14px;
-    width: 14px;
-    &:after {
-      margin-top: 2px;
-      margin-left: -4px;
-    }
-  }
-
-  .likes-count {
-    position: absolute;
-    left: -16px;
-    width: 14px;
-    font-size: 12px;
-    text-align: right;
-  }
-
-  .has-my-like {
-    color: #e66;
-  }
-}
-
-.actions-overlay {
-  position: fixed;
-  top: 0;
-  bottom: 0;
-  left: 0;
-  right: 0;
-  background-color: rgba(153,153,153, 0.7);
-}
-
-.actions-panel {
-  position: absolute;
-  bottom: 0;
-  left: 0;
-  right: 0;
-  background-color: #dddddd;
-
-  .likes-panel {
-    display: flex;
-
-    &.big {
-      height: 190px;
-    }
-
-    .arrow {
-      width: 35px;
-      font-size: 22px;
-      line-height: 40px;
-      text-align: center;
-    }
-
-    .likes {
-      flex: 1;
-      line-height: 40px;
-      font-size: 16px;
-      margin: 0;
-      padding-right: 15px;
-    }
-  }
-
-  .comment-likes-list {
-    padding-top: 8px;
-    line-height: 24px;
-  }
-
-  .mention-actions {
-    width: 100%;
-    background-color: white;
-    padding: 12px 0;
-  }
-
-  .mention-action {
-    height: 42px;
-    font-size: 16px;
-    line-height: 42px;
-    width: 100%;
-    padding: 0;
-    text-align: left;
-    border: none;
-    outline: none;
-    background-color: white;
-    &:active {
-      background-color: #ffb914;
-      .fa-heart {
-        color: #e66;
-      }
-    }
-   .fa {
-      position: relative;
-      top: -1px;
-      font-size: 14px;
-      width: 35px;
-      text-align: center;
-    }
-  }
-}
