+++ styles/shared/comment-likes.scss
@@ -1,13 +1,57 @@
-.comment-likes {
+.comment-likes-container {
   position: absolute;
-  top: 2px;
-  left: -34px;
-  display: none;
+  left: -15px;
+  width: 34px;
+  height: 18px;
+  display: flex;
+  justify-content: flex-end;
+  padding-right: 2px;
+  @media (max-width: 768px) {
+    .comment-likes.has-my-like + .comment-time, .comment-likes.liked + .comment-time {
+      display: none;
+    }
+    .comment-likes:not(.has-my-like, .liked) {
+      display: none;
+    }
+  }
+}
 
-  @media (min-width: 1025px) {
-    display: flex;
+.comment-time {
+  margin-right: 2px;
+}
+
+.comment-icon {
+  &, &:hover {
+    text-decoration: none;
   }
 
+  &.feed-comment-dot {
+    height: 8px;
+    width: 14px;
+  }
+
+  &.fa-stack {
+    width: 14px;
+    margin-top: 1px;
+    margin-right: 6px;
+    line-height: 1em;
+    padding-top: 4px;
+  }
+  .fa-plus, .fa-square {
+    font-size: 60%;
+    top: 7px;
+    left: 4px;
+  }
+}
+
+.comment-likes {
+  display: inline-flex;
+  align-items: flex-end;
+  margin-right: 2px;
+  margin-top: 1px;
+  @media (min-width: 769px) {
+    margin-right: 6px;
+  }
   &:not(.liked) {
     opacity: 0;
     transition: opacity .25s;
@@ -17,7 +61,7 @@
     position: relative;
   }
   .comment-heart:hover .icon {
-    transform: scale(1.5);
+    transform: scale(1.3);
     transition: transform .25s;
   }
 
@@ -37,8 +81,13 @@
   }
 
   .comment-count {
+    min-width: 15px;
+    text-align: right;
     cursor: pointer;
-    margin-right: 6px;
+    margin-right: 2px;
+    @media (min-width: 769px) {
+      margin-right: 6px;
+    }
     font-size: 12px;
     width: 7px;
     color: #898989;
@@ -51,18 +100,6 @@
     .icon, .comment-count {
       color: #e66;
     }
-  }
-
-  .comment-like-hint {
-    position: absolute;
-    top: -15px;
-    opacity: 0;
-    transition: opacity .3 s;
-  }
-
-  &:hover .comment-like-hint {
-    opacity: 1;
-    transition: opacity .3 s;
   }
 
   .comment-likes-list {
@@ -83,4 +120,88 @@
 .comment:hover .comment-likes:not(.liked) {
   opacity: 1;
   transition: opacity .25s;
+}
+
+.actions-overlay {
+  position: fixed;
+  top: 0;
+  bottom: 0;
+  left: 0;
+  right: 0;
+  background-color: rgba(153, 153, 153, 0.7);
+  .container {
+    position: absolute;
+    bottom: 0;
+    left: 0;
+    right: 0;
+  }
+}
+
+.actions-panel {
+  margin-left: -15px;
+  margin-right: -15px;
+  background-color: #dddddd;
+
+  @media (min-width: 768px) {
+    margin-left: 63px;
+  }
+
+  .likes-panel {
+    display: flex;
+
+    &.big {
+      height: 190px;
+    }
+
+    .arrow {
+      width: 35px;
+      font-size: 22px;
+      line-height: 40px;
+      text-align: center;
+    }
+
+    .likes {
+      flex: 1;
+      line-height: 40px;
+      font-size: 16px;
+      margin: 0;
+      padding-right: 15px;
+    }
+  }
+
+  .comment-likes-list {
+    padding-top: 8px;
+    line-height: 24px;
+  }
+
+  .mention-actions {
+    width: 100%;
+    background-color: white;
+    padding: 12px 0;
+  }
+
+  .mention-action {
+    height: 42px;
+    font-size: 16px;
+    line-height: 42px;
+    width: 100%;
+    padding: 0;
+    text-align: left;
+    border: none;
+    outline: none;
+    background-color: white;
+    &:active {
+      background-color: #ffb914;
+      .fa-heart {
+        color: #e66;
+      }
+    }
+    .fa {
+      position: relative;
+      top: -1px;
+      font-size: 14px;
+      width: 35px;
+      text-align: center;
+    }
+  }
 }
